fix(login): rename types file
